Update the readme
There is a point in time where one must just drop a conversation with someone who is using your code, clearly not familiar with networking, to try to shoe horn it into their own shit. And then they ask questions that make me question my sanity, or the very small amounts of whats left of it. Might as well put a TED talk in the readme so people know that this transport is NOT designed for non-Mirror usage and implementation.