Bugfixes: Fix #83, NREs emitted due to ClientDisconnect jank
ClientDisconnect seems to get called now when closing HostClient mode which originally set ClientState to Disconnecting... but in HostClient, the Client is not connected to the server so the main update loop tries to pump but gets null reference exceptions because some references are null (due to the reason it's not connected to the server). HostClients never connect to localhost themselves, they use a emulated connection. In theory, this patch should be fine, works with the latest Mirror Basic scene so uhh... I guess we're done here.